# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Joybox API for Library",
        "contact": {
            "name": "Support Team",
            "email": "giarsyani.nuli@gmail.com"
        },
        "version": "1.0"
    },
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Sign in",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "description": "User auth",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "customer@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "111111"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login Successful"
                    },
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Incorrect Details. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/books": {
            "get": {
                "tags": [
                    "Book"
                ],
                "summary": "Get list of book",
                "description": "Returns list of book",
                "operationId": "getBookList",
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Title of book",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Software Engineering"
                    },
                    {
                        "name": "author",
                        "in": "query",
                        "description": "Author of book",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "cover_id",
                        "in": "query",
                        "description": "Cover id of book",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Book"
                ],
                "summary": "Borrow and pick up book",
                "description": "Returns list of book",
                "operationId": "PickupBook",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "cover_id": {
                                        "type": "array",
                                        "items": {
                                            "type": "number",
                                            "example": "3956527"
                                        }
                                    },
                                    "pick_up_date": {
                                        "type": "string",
                                        "example": "2022-10-12 08:30"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "The fields are required"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/order": {
            "get": {
                "tags": [
                    "Order"
                ],
                "summary": "Get list of pickup book",
                "description": "Returns list of pickup book",
                "operationId": "getPickUpList",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Status of pickup",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "pending"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/order_status": {
            "post": {
                "tags": [
                    "Order"
                ],
                "summary": "Update pickup status",
                "operationId": "UpdatePickUpStatus",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "order_id": {
                                        "description": "order_id get from list pickup/order",
                                        "type": "string",
                                        "example": "1"
                                    },
                                    "status": {
                                        "description": "status of pickup",
                                        "type": "borrow",
                                        "example": "borrow"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "The fields are required"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Authentication",
            "description": "API Endpoints of Authentication"
        },
        {
            "name": "Book",
            "description": "API Endpoints of Book"
        },
        {
            "name": "Order",
            "description": "API Endpoints of PickUp Book"
        }
    ]
}